refactor(threads): split views and modal into separate files
